# perf(config_validation): make regex a global variable and avoid `[]byte` conversion