**What this PR does / why we need it**:

We can use `(*regexp.Regexp).MatchString` to avoid an unnecessary byte slice conversion and reduce allocation.

Benchmark code using `(*regexp.Regexp).MatchString`:
```go
func matchString(flagValue string) (string, error) {
	re := regexp.MustCompile(`^[a-z0-9]([-a-z0-9]*[a-z0-9])?(\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*\/[A-Za-z0-9\/\-._~%!$&'()*+,;=:]+$`)
	if !re.MatchString(flagValue) {
		return "", errors.New("the expected format is example.com/controller-name")
	}
	return flagValue, nil
}

func match(flagValue string) (string, error) {
	re := regexp.MustCompile(`^[a-z0-9]([-a-z0-9]*[a-z0-9])?(\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*\/[A-Za-z0-9\/\-._~%!$&'()*+,;=:]+$`)
	if !re.Match([]byte(flagValue)) {
		return "", errors.New("the expected format is example.com/controller-name")
	}
	return flagValue, nil
}

func BenchmarkGatewayAPIControllerNameFromFlagValueMatchString(b *testing.B) {
	for i := 0; i < b.N; i++ {
		matchString("example.com/controller-name")
	}
}
func BenchmarkGatewayAPIControllerNameFromFlagValueMatch(b *testing.B) {
	for i := 0; i < b.N; i++ {
		match("example.com/controller-name")
	}
}
```

Result:
```
goos: linux
goarch: amd64
pkg: github.com/kong/kubernetes-ingress-controller/v2/internal/manager
cpu: AMD Ryzen 7 PRO 4750U with Radeon Graphics
BenchmarkGatewayAPIControllerNameFromFlagValueMatchString-16    	   52533	     26469 ns/op	    9781 B/op	      78 allocs/op
BenchmarkGatewayAPIControllerNameFromFlagValueMatch-16          	   43731	     27882 ns/op	    9845 B/op	      79 allocs/op
```

---

Benchmark code using `(*regexp.Regexp).MatchString` and global regex:
```go
var gatewayAPIControllerNameRegex = regexp.MustCompile(`^[a-z0-9]([-a-z0-9]*[a-z0-9])?(\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*\/[A-Za-z0-9\/\-._~%!$&'()*+,;=:]+$`)

func matchString(flagValue string) (string, error) {
	if !gatewayAPIControllerNameRegex.MatchString(flagValue) {
		return "", errors.New("the expected format is example.com/controller-name")
	}
	return flagValue, nil
}

func match(flagValue string) (string, error) {
	if !gatewayAPIControllerNameRegex.Match([]byte(flagValue)) {
		return "", errors.New("the expected format is example.com/controller-name")
	}
	return flagValue, nil
}

func BenchmarkGatewayAPIControllerNameFromFlagValueMatchString(b *testing.B) {
	for i := 0; i < b.N; i++ {
		matchString("example.com/controller-name")
	}
}
func BenchmarkGatewayAPIControllerNameFromFlagValueMatch(b *testing.B) {
	for i := 0; i < b.N; i++ {
		match("example.com/controller-name")
	}
}
```

Result:
```
goos: linux
goarch: amd64
pkg: github.com/kong/kubernetes-ingress-controller/v2/internal/manager
cpu: AMD Ryzen 7 PRO 4750U with Radeon Graphics
BenchmarkGatewayAPIControllerNameFromFlagValueMatchString-16    	 2052745	       557.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkGatewayAPIControllerNameFromFlagValueMatch-16          	  920378	      1233 ns/op	      32 B/op	       1 allocs/op
```

**Which issue this PR fixes**:

**Special notes for your reviewer**:


**PR Readiness Checklist**:

Complete these before marking the PR as `ready to review`:

- [x] the `CHANGELOG.md` release notes have been updated to reflect any significant (and particularly user-facing) changes introduced by this PR